+++ src/main/java/kotlinx/reflect/lite/impl/FunctionMetadataImpl.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2010-2017 JetBrains s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package kotlinx.reflect.lite.impl
+
+import kotlinx.reflect.lite.FunctionMetadata
+import kotlinx.reflect.lite.ParameterMetadata
+import kotlinx.reflect.lite.TypeMetadata
+import org.jetbrains.kotlin.serialization.ProtoBuf
+import org.jetbrains.kotlin.serialization.deserialization.NameResolver
+
+internal class FunctionMetadataImpl(
+        private val proto: ProtoBuf.Function,
+        private val nameResolver: NameResolver
+) : CallableMetadataImpl(), FunctionMetadata {
+    override val parameters: List<ParameterMetadata>
+        get() = proto.valueParameterList.map { ParameterMetadataImpl(it, nameResolver) }
+
+    override val returnType: TypeMetadata
+        get() = TypeMetadataImpl(proto.returnType, nameResolver)
+}
